+++ engine/lite/embeddedDeploymentManager/src/test/scala/pl/touk/nussknacker/streaming/embedded/RequestResponseEmbeddedDeploymentManagerTest.scala
@@ -87,7 +87,7 @@ class RequestResponseEmbeddedDeploymentManagerTest extends AnyFunSuite with Matc
     }
 
     request.body("""{ productId: 15 }""").send().body shouldBe Right("""{"transformed":15}""")
-    request.body("""Not a correct json""").send().body shouldBe Left("""[{"message":"A JSONObject text must begin with '{' at 1 [character 2 line 1]","nodeId":"source"}]""")
+    request.body("""Not a correct json""").send().body shouldBe Left("""[{"message":"#: expected type: JSONObject, found: String","nodeId":"source"}]""")
     request.body("""{ productId: "11"}""").send().body shouldBe Left("""[{"message":"#/productId: expected type: Integer, found: String","nodeId":"source"}]""")
 
     manager.cancel(name, User("a", "b")).futureValue